update API and UserController

+++ app/Http/Controllers/API/UserController.php
@@ -7,85 +7,100 @@ use App\Models\User;
 use Illuminate\Database\QueryException;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
-use Illuminate\Support\Facades\Hash;
+use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Session;
+use Illuminate\Support\Facades\Storage;
 
 class UserController extends Controller
 {
-    //
-    public function register(Request $request){
-        try{
-            $user = new User();
-            $user->name = $request->name;
-            $user->email = $request->email;
-            $user->password = Hash::make($request->password);
-            $user->save();
 
-            $success = true;
+    public function list() {
+        try {
+            //code...
+            if (Auth::check()) {
+
+                $users = User::all()->toArray();
+                $response = [
+                    'success' => true,
+                    'users' => $users,
+                ];
+                return response()->json($response, 200);
+            }
+            $response = [
+                'success' => false,
+                'message' => 'Usuário não autenticado',
+            ];
+            return response()->json($response, 401);
+        } catch (\Throwable $th) {
+            $response = [
+                'success' => false,
+                'message' => $th->getMessage()
+            ];
+            return response()->json($response, 500);
+        }
+    }
+    public function show($id) {
+        try {
+            //code...
+            if (Auth::check()) {
+
+                $user = User::find($id);
+                $response = [
+                    'success' => true,
+                    'users' => $user,
+                ];
+                return response()->json($response, 200);
+            }
+            $response = [
+                'success' => false,
+                'message' => 'Usuário não autenticado',
+            ];
+            return response()->json($response, 401);
+        } catch (\Throwable $th) {
+            $response = [
+                'success' => false,
+                'message' => $th->getMessage()
+            ];
+            return response()->json($response, 500);
+        }
+    }
+    public function register(Request $request)
+    {
+        // validate the request data
+
+        $request->validate([
+            'name' => 'required|string|max:255',
+            'email' => 'required|string|email|max:255|unique:users',
+            'password' => 'required|string|min:8',
+            'phone' => 'nullable|string|digits_between:11,11'
+        ]);
+
+        try {
+            //code...
+            // create the new user in the database
+            $user = User::create([
+                'status' => 'active',
+                'name' => $request->name,
+                'email' => $request->email,
+                'password' => bcrypt($request->password),
+                'phone' => $request->phone ?? null,
+            ]);
+
+            // log the user in
+            Auth::login($user, true);
             $message = 'Usuário registrado com sucesso';
             $response = [
-                'success' => $success,
-                'message' => $message,
+                'user'      => Auth::user()->name,
+                'userId'    => Auth::user()->id,
+                'token'     => $user->createToken("API TOKEN")->plainTextToken,
+                'success'   => true,
+                'message'   => 'Usuário registrado e logado com sucesso',
             ];
             return response()->json($response, 200);
-        } catch(QueryException $error){
+        } catch (QueryException $error) {
             $success = false;
             $message = $error;
 
-        }
-        $response = [
-            'success' => $success,
-            'message' => $message,
-        ];
-        return response()->json($response, 500);
-
-    }
-
-    public function login(Request $request){
-
-
-        $response = [];
-        $dados = [
-            'email' => $request->email,
-            'password' => $request->password
-        ];
-
-        if(Auth::attempt($dados)){
-            $response = [
-                'user'    => Auth::user()->name,
-                'userId'  => Auth::user()->id,
-                'success' => true,
-                'message' => "Usuario logado com sucesso"
-            ];
-            return response()->json($response, 200);
-        }else{
-
-            $success = false;
-            $message = 'email ou login inválidos';
-        }
-        $response = [
-            'success' => $success,
-            'message' => $message
-        ];
-        return response()->json($response, 400);
-
-    }
-
-    public function logout(){
-
-        $response = [];
-        try{
-            Session::flush();
-            $success = true;
-            $message = 'Usuario deslogado com sucesso';
-            $response = [
-                'success' => $success,
-                'message' => $message,
-            ];
-            return response()->json($response, 200);
-        } catch(QueryException $error){
-            $success = false;
-            $message = $error->getMessage();
             $response = [
                 'success' => $success,
                 'message' => $message,
@@ -94,20 +109,102 @@ class UserController extends Controller
         }
 
     }
-    public function update(Request $request){
 
+    public function login(Request $request)
+    {
+        //validate the request data
         try {
-            $user = User::find(Auth::user()->id);
-            $user->update($request->all());
-            $this->logout();
+            //code...
+            $request->validate([
+                'email' => 'required|string|email|max:255',
+                'password' => 'required|string',
+            ]);
+
+            //attempt to authenticate the user
+            $credentials = $request->only('email', 'password');
+            if (Auth::attempt($credentials, true)) {
+                $user = User::where('email', $credentials['email'])->first();
+                Auth::login($user, true);
+                $response = [
+                    'user'      => Auth::user()->name,
+                    'userId'    => $user->id,
+                    'token'     => $user->createToken("API TOKEN")->plainTextToken,
+                    'success'   => true,
+                    'message'   => "Usuário logado com sucesso"
+                ];
+                return response()->json($response, 200);
+            } else {
+                //if the user's credentials are invalid, redirect back to the login page with an error message
+                $response = [
+                    'success' => false,
+                    'message' => 'email ou senha inválidos'
+                ];
+                return response()->json($response, 401);
+            }
+        } catch (\Throwable $th) {
+            $response = [
+                'success' => false,
+                'message' => $th->getMessage()
+            ];
+            return response()->json($response, 500);
+        }
+
+    }
+
+
+    public function update(Request $request) {
+        try {
+            //code...
+            if(Auth::user()) {
+
+                $user = Auth::user();
+                $request->validate([
+                    'name' => 'sometimes|string|max:255',
+                    'email' => 'sometimes|string|email|max:255|unique:users,email,'.$user->id,
+                    'password' => 'sometimes|string|min:8|confirmed',
+                    'phone' => 'nullable|string|digits_between:11,11',
+                ]);
+
+
+                $user->fill($request->only([
+                    'email'
+                ]));
+
+                if ($request->filled('password')) {
+                    $user->password = bcrypt($request->password);
+                }
+                $user->save();
+                $id = $user->id;
+
+                if ($request->hasFile('userImage')) {
+                    $image = $request->file('userImage');
+                    if ($image->isValid()) {
+
+                        $file_name = $image->getFilename();
+                        $image->store('images', 'storage');
+                        DB::transaction(function () use( $id , $file_name){
+                            DB::insert('insert into user_images (user_id, user_name_image) values (?, ?)', [$id, $file_name]);
+
+                        }, 5);
+                    }
+                }
+                Session::flush();
+
+                $response = [
+                    'success' => true,
+                    'message' => "Usuário atualizado com sucesso",
+                ];
+                return response()->json($response, 200);
+            }
 
             $response = [
-                'success' => true,
-                'message' => "Usuário atualizado com sucesso",
+                'success' => false,
+                'message' => "Usuário não autenticado",
             ];
-            return response()->json($response, 200);
-        } catch (\Throwable $th) {
+            return response()->json($response, 401);
 
+        } catch (\Throwable $th) {
+            //throw $th;
             $response = [
                 'success' => false,
                 'message' => $th->getMessage(),
@@ -117,31 +214,29 @@ class UserController extends Controller
 
 
     }
-    public function list(){
 
+    public function logout() {
         try {
+            //code...
 
-            $users = User::all();
-            return response()->json($users, 200);
-
-        } catch (\Throwable $th) {
-
+            if(Auth::check()){
+                Auth::user()->tokens()->delete();
+                $response = [
+                    'success' => true,
+                    'message' => 'Usuário deslogado com sucesso',
+                ];
+                return response()->json($response, 200);
+            }
             $response = [
                 'success' => false,
-                'message' => $th->getMessage(),
+                'message' => "Usuário não autenticado",
             ];
-            return response()->json($response, 500);
-        }
-    }
-    public function show($id){
+            return response()->json($response, 401);
 
-        try {
 
-            $user = User::select(['name','email','phone','photo'])->find($id);
-            return response()->json($user, 200);
 
         } catch (\Throwable $th) {
-
+            //throw $th;
             $response = [
                 'success' => false,
                 'message' => $th->getMessage(),
@@ -151,24 +246,99 @@ class UserController extends Controller
 
 
     }
-    public function remove($id){
-        $user = User::find($id);
 
+    public function delete($id) {
         try {
-            $user->delete();
+            //code...
+
+            if(Auth::user()) {
+
+                $user = User::find($id);
+                $user->delete();
+                $response = [
+                    'success' => true,
+                    'message' => 'Usuário removido com sucesso',
+                ];
+                return response()->json($response, 200);
+            }
             $response = [
-                'success' => true,
-                'message' => "Usuário removido com sucesso",
+                'success' => false,
+                'message' => "Usuário não autenticado",
             ];
-            return response()->json($response, 200);
+            return response()->json($response, 401);
 
         } catch (\Throwable $th) {
+            //throw $th;
             $response = [
                 'success' => false,
                 'message' => $th->getMessage(),
             ];
             return response()->json($response, 500);
         }
+    }
 
+    public function imageUser(Request $request) {
+        try {
+            //code...
+
+            if(Auth::check()) {
+
+                $id = Auth::user()->id;
+                if ($id && $request->hasFile('userImage')) {
+
+                    $image = $request->file('userImage');
+                    if ($image->isValid()) {
+
+                        $file_name = $image->getClientOriginalName();
+                        $path = Storage::disk('local')->putFileAs('images', $image, $file_name);
+                        
+                        DB::transaction(function () use( $id , $file_name){
+                            $result = DB::select('select id from user_images where user_id = :id', ['id' => $id]);
+
+                            if(!$result) {
+                                DB::insert('insert into user_images (user_id, user_name_image) values (?, ?)', [$id, $file_name]);
+                                return;
+                            }
+                            DB::update(
+                                'update user_images
+                                    set user_name_image = ?
+                                    where user_id = ?',[$file_name, $id]);
+
+
+                        }, 5);
+                        $response = [
+                            'success' => true,
+                            'imagemPath' =>  $path,
+                            'message' => 'Imagem enviada com sucesso!'
+                        ];
+                        return response()->json($response, 200);
+                    }
+                    $response = [
+                        'success' => false,
+                        'message' => "O arquivo enviado não é valido!",
+                    ];
+                    return response()->json($response, 400);
+                }
+                $response = [
+                    'success' => false,
+                    'message' => "Arquivo não encontrado!",
+                ];
+                return response()->json($response, 400);
+            }
+            $response = [
+                'success' => false,
+                'message' => "Usuário não autenticado",
+            ];
+            return response()->json($response, 401);
+
+        } catch (\Throwable $th) {
+            //throw $th;
+            $response = [
+                'success' => false,
+                'message' => $th->getMessage(),
+            ];
+            return response()->json($response, 500);
+        }
     }
 }
+
